backend/requirements.txt
# Dependencies for career prediction backend# Core LangGraph and LangChain dependencies
langgraph==0.6.7
langchain==0.3.27
langchain-openai==0.3.33
langchain-core==0.3.76
langchain-community==0.3.29
langgraph-supervisor==0.0.29

# LangGraph sub-dependencies
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6

# Monitoring and debugging
langsmith==0.4.28

# Distributed Tracing with OpenTelemetry and Jaeger
# Latest stable versions as of October 2025
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-langchain==0.47.3
opentelemetry-instrumentation-fastapi==0.59b0

# Web search and data retrieval
langchain-tavily==0.2.11
duckduckgo-search==8.1.1
ddgs
requests==2.32.5
requests-toolbelt==1.0.0

# OpenAI and related
openai==1.107.3
tiktoken==0.11.0

# Google Gemini AI
langchain-google-genai==2.0.10
google-generativeai==0.8.4

# Environment and configuration
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1

# Data processing and analysis
pandas==2.3.2
numpy==2.3.3
python-dateutil==2.9.0.post0

# Web API Framework
fastapi==0.116.1
starlette==0.47.3
uvicorn[standard]==0.35.0
websockets==15.0.1

# HTTP clients and tools
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
aiohttp==3.12.15
httpx-sse==0.4.1

# CORS support
fastapi-cors==0.0.6
environs==14.3.0

# Database
SQLAlchemy==2.0.43
greenlet==3.2.4

# JSON and serialization
orjson==3.11.3
ormsgpack==1.10.0
dataclasses-json==0.6.7
marshmallow==3.26.1

# Text processing
langchain-text-splitters==0.3.11
PyYAML==6.0.2
regex==2025.9.1

# RAG System - Document processing and vector storage
chromadb==0.5.23
pypdf==5.1.0
pymupdf==1.26.6
python-docx==1.1.0

# Data Ingestion System
watchdog==4.0.0              # File system monitoring
APScheduler==3.10.4          # Scheduled tasks and job scheduling
beautifulsoup4==4.12.3       # Web scraping and HTML parsing
selenium==4.16.0             # Browser automation for JS-heavy sites
aiofiles==23.2.1             # Async file operations
aiosqlite==0.19.0            # Async SQLite for ingestion registry
prometheus-client==0.19.0    # Metrics collection and export
lxml==5.3.0                  # XML/HTML parser for web scraping

# Async and networking
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
multidict==6.6.4
yarl==1.20.1
frozenlist==1.7.0
propcache==0.3.2

# Compression and hashing
zstandard==0.25.0
xxhash==3.5.0
brotli==1.1.0
zopfli==0.2.3.post1

# Utilities
attrs==25.3.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tenacity==9.1.2
tqdm==4.67.1
click==8.2.1
colorama==0.4.6
watchfiles==1.1.0

# JSON processing
jsonpatch==1.33
jsonpointer==3.0.0
jiter==0.11.0

# Time and date
pytz==2025.2
tzdata==2025.2
six==1.17.0

# Security and certificates
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
distro==1.9.0

# Development tools
mypy-extensions==1.1.0
packaging==25.0

# Testing
pytest==8.4.2
pytest-asyncio==1.2.0
pluggy==1.6.0
iniconfig==2.1.0
pygments==2.19.2

# Report generation
jinja2==3.1.6
MarkupSafe==3.0.2
weasyprint==66.0
pydyf==0.11.0
cffi==2.0.0
pycparser==2.23
tinyhtml5==2.0.0
tinycss2==1.4.0
cssselect2==0.8.0
webencodings==0.5.1
pyphen==0.17.2
pillow==11.3.0
fonttools==4.59.2